Optimize bitset boundary conversion and document results
- Optimization A: remove redundant uint8 copy in bitset_to_bool_array (bitset_backend.py).
- Add CM_bitset_boundary_report.md with:
  - audit table + code-truth boundary path description (hybrid vs partial_hybrid)
  - new instrumentation fields list
  - pre/post benchmark medians and boundary breakdown
  - verdict on whether boundary dominates hybrid slowdown

Benchmarks (threshold=7):
- Pre:  bench_runs/bench_boundary_pre_t7_{raw,summary}.csv
- Post: bench_runs/bench_boundary_post_t7_{raw,summary}.csv
Representative outcome (medians on this slice): boundary_time reduced ~1.4–2.0× for n<=12, with total-time improvements varying by n; verdict: boundary overhead is real but not the dominant bottleneck.

Validation: python -m pytest -q (30 passed); cm_bench compare run succeeded.

+++ bitset_backend.py
@@ -80,7 +80,7 @@ def bitset_to_bool_array(bits: int, n_vars: int) -> np.ndarray:
     n_bytes = (n_rows + 7) // 8
     raw = bits.to_bytes(n_bytes, byteorder="little", signed=False)
     unpacked = np.unpackbits(np.frombuffer(raw, dtype=np.uint8), bitorder="little")
-    return unpacked[:n_rows].astype(np.uint8)
+    return unpacked[:n_rows]
 
 
 def bitset_to_bool_hypercube(bits: int, n_vars: int) -> np.ndarray:
